+++ relay-server/src/services/processor.rs
@@ -70,6 +70,7 @@ use relay_threading::AsyncPool;
 #[cfg(feature = "processing")]
 use {
     crate::services::global_rate_limits::{GlobalRateLimits, GlobalRateLimitsServiceHandle},
+    crate::services::processor::nnswitch::SwitchProcessingError,
     crate::services::store::{Store, StoreEnvelope},
     crate::utils::{Enforcement, ItemAction},
     itertools::Itertools,
@@ -103,6 +104,9 @@ mod playstation;
 mod standalone;
 #[cfg(feature = "processing")]
 mod unreal;
+
+#[cfg(feature = "processing")]
+mod nnswitch;
 
 /// Creates the block only if used with `processing` feature.
 ///
@@ -529,6 +533,10 @@ pub enum ProcessingError {
 
     #[error("replay filtered with reason: {0:?}")]
     ReplayFiltered(FilterStatKey),
+
+    #[cfg(feature = "processing")]
+    #[error("nintendo switch dying message processing failed {0:?}")]
+    InvalidNintendoDyingMessage(#[source] SwitchProcessingError),
 }
 
 impl ProcessingError {
@@ -550,6 +558,9 @@ impl ProcessingError {
             Self::InvalidTimestamp => Some(Outcome::Invalid(DiscardReason::Timestamp)),
             Self::DuplicateItem(_) => Some(Outcome::Invalid(DiscardReason::DuplicateItem)),
             Self::NoEventPayload => Some(Outcome::Invalid(DiscardReason::NoEventPayload)),
+
+            #[cfg(feature = "processing")]
+            Self::InvalidNintendoDyingMessage(_) => Some(Outcome::Invalid(DiscardReason::Payload)),
 
             // Processing-only outcomes (Sentry-internal Relays)
             #[cfg(feature = "processing")]
@@ -1574,6 +1585,7 @@ impl EnvelopeProcessorService {
         if_processing!(self.inner.config, {
             unreal::expand(managed_envelope, &self.inner.config)?;
             playstation::expand(managed_envelope, &project_info)?;
+            nnswitch::expand(managed_envelope)?;
         });
 
         let extraction_result = event::extract(
@@ -2239,6 +2251,7 @@ impl EnvelopeProcessorService {
                             extracted_metrics: extracted_metrics.map_or(ProcessingExtractedMetrics::new(), |e| e)
                         }),
                         Err(error) => {
+                            relay_log::trace!("Executing {fn} failed: {error}", fn = stringify!($fn_name), error = error);
                             if let Some(outcome) = error.to_outcome() {
                                 managed_envelope.reject(outcome);
                             }